+++ src/entrypoint.c
@@ -1,8 +1,8 @@
 // We need to forward routine registration from C to Rust
 // to avoid the linker removing the static library.
 
-void R_init_rpolars_extendr(void *dll);
+void R_init_polars_extendr(void *dll);
 
-void R_init_rpolars(void *dll) {
-    R_init_rpolars_extendr(dll);  //first or last
+void R_init_polars(void *dll) {
+    R_init_polars_extendr(dll);  //first or last
 }