+++ src/rust/src/rdatatype.rs
@@ -289,7 +289,7 @@ pub fn new_join_type(s: &str) -> pl::JoinType {
         "outer" => pl::JoinType::Outer,
         "semi" => pl::JoinType::Semi,
         "anti" => pl::JoinType::Anti,
-        _ => panic!("rpolars internal error: jointype not recognized"),
+        _ => panic!("polars internal error: jointype not recognized"),
     }
 }