+++ tests/testthat/test-minipolars_options.R
@@ -1,6 +1,6 @@
 test_that("get options", {
-  opts = pl$get_rpolars_options()
-  expect_equal(opts,as.list(rpolars:::rpolars_optenv))
+  opts = pl$get_polars_options()
+  expect_equal(opts,as.list(polars:::polars_optenv))
 })
 
 
@@ -8,16 +8,16 @@ test_that("get options", {
 test_that("set wrong options", {
 
   #pass a not defined options
-  expect_error(pl$set_rpolars_options(undefined_opt = 42 , undefined_opt_2 = 214))
+  expect_error(pl$set_polars_options(undefined_opt = 42 , undefined_opt_2 = 214))
 
   #pass a not named option
-  expect_error(pl$set_rpolars_options(42))
+  expect_error(pl$set_polars_options(42))
 
   #existing option but wrong type
-  expect_error(pl$set_rpolars_options(strictly_immutable = 42))
+  expect_error(pl$set_polars_options(strictly_immutable = 42))
 
   #existing option but wrong length
-  expect_error(pl$set_rpolars_options(strictly_immutable = c(TRUE,TRUE)))
+  expect_error(pl$set_polars_options(strictly_immutable = c(TRUE,TRUE)))
 
 })
 
@@ -31,13 +31,13 @@ test_that("set/get/reset an option", {
   expect_true(all(names(df)!=names(df_immutable_copy)))
 
   #check current state
-  before_opt = pl$get_rpolars_options()$strictly_immutable
+  before_opt = pl$get_polars_options()$strictly_immutable
 
   #setting and option returns the previous/state state as defualt
-  before_opt2 = pl$set_rpolars_options(strictly_immutable= !before_opt)[[1]]
+  before_opt2 = pl$set_polars_options(strictly_immutable= !before_opt)[[1]]
 
   #get new state
-  after_opt = pl$get_rpolars_options()$strictly_immutable
+  after_opt = pl$get_polars_options()$strictly_immutable
 
   #check change took place
   expect_equal(before_opt, !after_opt)
@@ -51,7 +51,8 @@ test_that("set/get/reset an option", {
 
   #check returning to default state, if fail, maybe testthat did not start in default
   #state which is an error.
-  pl$reset_rpolars_options()
-  default_opt = pl$get_rpolars_options()$strictly_immutable
+  pl$reset_polars_options()
+  default_opt = pl$get_polars_options()$strictly_immutable
   expect_equal(before_opt,  default_opt)
 })
+